Mutex + Midi Polling
Added a mutex since we modify data in 2 threads in the midi driver.
Also, we constantly poll in the driver, based on the timer expireing,
but we just return if there is no midi input to work with.  Before..
midi input polling didn't happen if you didn't connect to a midi input
through the commandline.